+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## 4.0.0 - Unreleased
+## 4.0.0 - 2026-07-11
 
 ### Breaking Changes